+++ package/MDAnalysis/coordinates/__init__.py
@@ -733,7 +733,6 @@ from . import INPCRD
 from . import LAMMPS
 from . import MOL2
 from . import OpenMM
-from . import ParmEd
 from . import PDB
 from . import PDBQT
 from . import PQR
@@ -750,4 +749,3 @@ from . import GSD
 from . import null
 from . import NAMDBIN
 from . import FHIAIMS
-from . import RDKit
