+++ package/doc/sphinx/source/documentation_pages/converters/ParmEd.rst
@@ -1,0 +1,3 @@
+.. automodule:: MDAnalysis.converters.ParmEdParser
+
+.. automodule:: MDAnalysis.converters.ParmEd
