--- package/doc/sphinx/source/documentation_pages/converters/ParmEdParser.rst
@@ -1,3 +1,0 @@
-.. automodule:: MDAnalysis.topology.ParmEdParser
-
-.. automodule:: MDAnalysis.coordinates.ParmEd
