+++ package/doc/sphinx/source/documentation_pages/converters/RDKit.rst
@@ -1,0 +1,3 @@
+.. automodule:: MDAnalysis.converters.RDKitParser
+
+.. automodule:: MDAnalysis.converters.RDKit
