--- package/doc/sphinx/source/documentation_pages/converters/RDKitParser.rst
@@ -1,3 +1,0 @@
-.. automodule:: MDAnalysis.topology.RDKitParser
-
-.. automodule:: MDAnalysis.coordinates.RDKit
