+++ package/doc/sphinx/source/documentation_pages/converters/init.rst
@@ -1,0 +1,1 @@
+.. automodule:: MDAnalysis.converters.__init__
